Bug 102813 [Viewers] Text compare editor: vertical scroll synchronization does not work with large files

+++ bundles/org.eclipse.compare/compare/org/eclipse/compare/contentmergeviewer/TextMergeViewer.java
@@ -4938,7 +4938,7 @@ public class TextMergeViewer extends ContentMergeViewer implements IAdaptable  {
 				if (viewHeight <= 0) {
 					v= 0;
 				} else {
-					v= (v*viewHeight)/virtualHeight;
+					v= (int) (v * ((double)viewHeight/virtualHeight));
 				}
 				return viewPos+v;
 			}
